+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning].
 ## [Unreleased]
 
 - Fix default shell for windows when `%COMSPEC%` is missing.
+- Fix escaping of `~` for certain input strings with `escape` and `escapeAll`
+  when `interpolation` is set to `true` in the options.
 - Improve error messaging for certain non-string arguments.
 
 ## [1.5.0] - 2022-02-14