+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Versioning].
 
 ## [Unreleased]
 
+- _No changes yet_
+
+## [1.5.1] - 2022-02-19
+
 - Fix default shell for windows when `%COMSPEC%` is missing.
 - Fix escaping of `~` for certain input strings with `escape` and `escapeAll`
   when `interpolation` is set to `true` in the options.